Dom changes

+++ dom.js
@@ -6,22 +6,7 @@
   var container = document.getElementById('todo-container');
   var addTodoForm = document.getElementById('add-todo');
 
-  var state = [{
-      id: -3,
-      description: 'first todo',
-      done: false
-    },
-    {
-      id: -2,
-      description: 'second todo',
-      done: false
-    },
-    {
-      id: -1,
-      description: 'third todo',
-      done: false
-    },
-  ]; // this is our initial todoList
+  var state = []; // this is our initial todoList
 
   // This function takes a todo, it returns the DOM node representing that todo
   var createTodoNode = function (todo) {
@@ -46,21 +31,25 @@
 
     // this adds the delete button
     var deleteButtonNode = document.createElement('i');
-    deleteButtonNode.classList.add('fas', 'fa-trash-alt','deleteicon')
+    deleteButtonNode.classList.add('fas', 'fa-trash-alt', 'deleteicon')
     deleteButtonNode.addEventListener('click', function (event) {
-    
-      var newState = todoFunctions.deleteTodo(state, todo.id);
-      update(newState);
+      if (confirm('Are you sure ?')) {
+        var newState = todoFunctions.deleteTodo(state, todo.id);
+        update(newState);
+
+      }
+
     });
     todoNode.appendChild(deleteButtonNode);
 
     // add markTodo button
 
     var markButtonNode = document.createElement('i');
-    markButtonNode.classList.add("fas", "fa-check",'check_icon');
+    markButtonNode.classList.add("fas", "fa-check", 'check_icon');
     markButtonNode.addEventListener('click', function (event) {
       let marked = todoFunctions.markTodo(state, todo.id);
       let newState = todoFunctions.sortTodos(marked, arrange);
+      console.log(marked, newState);
       update(newState);
     });
     todoNode.appendChild(markButtonNode);
@@ -77,17 +66,13 @@
       // what is inside event.target?
       event.preventDefault();
       let description = document.querySelector('input[type="text"]').value;
-      console.log(description);
-      console.log(typeof description);
       // event.target ....
       if (!description.trim()) {
-        alert("No spaces allowed");
+        alert("Not allowed");
       }
       // hint: todoFunctions.addTodo
       else {
         let newState = todoFunctions.addTodo(state, description);
-        console.log(newState);
-        // ?? change this!
         update(newState);
       }
     });
@@ -96,7 +81,7 @@
 
   // Create Arrange Function 
 
-  var arrange = function(a,b){
+  var arrange = function (a, b) {
     return a.done - b.done;
   }
